+++ loader/wsi.h
@@ -23,7 +23,7 @@
 #pragma once
 
 #include "vk_loader_platform.h"
-#include "loader.h"
+#include "loader_common.h"
 
 typedef struct {
     union {
